wip: pin toolchain to 1.69.0
This is the toolchain version distributed with the current illumos from
the oxide helios-engvm as well as NixOS 23.05. This brings us inline w/
the offline-keystore as well as offline-keystore-os.

This probably won't work given the change to `dice-mfg/src/lib.rs`:
`std::path::absolute` is meant to be a drop in replacement for
`std::fs::canonicalize` but the reverse isn't the case. Our use of
`path::absolute` is to get the absolute path of a file that doesn't yet
exist. Doing so w/ `fs::canonicalize` will fail since the path doesn't
yet exist.

Might be possible to touch the file before canonicalizing the path, then
removing it after normalization. It's a hack but it'll probably work.