Update `verify_cert_chain` function to verify against multiple roots.
We expect that in the common case consumers will have more than one
trust anchor that they trust. The previous interface required that
consumers verify the cert chain against each root themselves. This works
but it's pretty inefficient since it causes signatures over the whole
chain to be verified for each root. This change makes this more
efficient by accepting multiple roots as an optional slice of
`Certificate` instances. Doing so required rethinking the base,
recursion and termination cases.